refactor(merge-tree): Clean up Perspective API
